backend/requirements.txt
fastapi==0.115.0
uvicorn==0.30.6
python-multipart==0.0.9
pydantic==2.8.2
pillow==10.4.0
paddleocr==2.7.3
paddlepaddle==2.6.1
sqlalchemy==2.0.32
pytest==8.3.3
